+++ src/app/[locale]/layout.tsx
@@ -1,10 +1,8 @@
 import "./globals.css";
-import { NextIntlClientProvider } from "next-intl";
-import { notFound } from "next/navigation";
-import { unstable_setRequestLocale } from "next-intl/server";
 import type { Metadata } from "next";
 import { Montserrat } from "next/font/google";
 import Header from "@/components/header";
+import Providers from "@/providers";
 
 const montserrat = Montserrat({ subsets: ["latin"] });
 
@@ -13,12 +11,6 @@ interface IRootLayout {
   params: {
     locale: string;
   };
-}
-
-const locales = ["en", "pl"];
-
-export function generateStaticParams() {
-  return locales.map((locale) => ({ locale }));
 }
 
 export const metadata: Metadata = {
@@ -30,28 +22,16 @@ export default async function RootLayout({
   children,
   params: { locale },
 }: IRootLayout) {
-  let messages;
-  try {
-    messages = (await import(`../../messages/${locale}.json`)).default;
-  } catch (error) {
-    notFound();
-  }
-
-  const isValidLocale = locales.some((cur) => cur === locale);
-  if (!isValidLocale) notFound();
-
-  unstable_setRequestLocale(locale);
-
   return (
-    <html lang={locale}>
-      <body
-        className={`${montserrat.className} bg-backgrounds-main flex justify-center items-center h-screen px-3`}
-      >
-        <NextIntlClientProvider locale={locale} messages={messages}>
+    <Providers locale={locale}>
+      <html lang={locale}>
+        <body
+          className={`${montserrat.className} bg-backgrounds-main flex justify-center items-center h-screen px-3`}
+        >
           <Header />
           {children}
-        </NextIntlClientProvider>
-      </body>
-    </html>
+        </body>
+      </html>
+    </Providers>
   );
 }
