+++ src/providers/NextIntlProvider.tsx
@@ -1,0 +1,37 @@
+import { NextIntlClientProvider } from "next-intl";
+import { unstable_setRequestLocale } from "next-intl/server";
+import { notFound } from "next/navigation";
+
+const locales = ["en", "pl"];
+
+export function generateStaticParams() {
+  return locales.map((locale) => ({ locale }));
+}
+
+const NextIntlProvider = async ({
+  children,
+  locale,
+}: {
+  children: React.ReactNode;
+  locale: string;
+}) => {
+  let messages;
+  try {
+    messages = (await import(`../messages/${locale}.json`)).default;
+  } catch (error) {
+    notFound();
+  }
+
+  const isValidLocale = locales.some((cur) => cur === locale);
+  if (!isValidLocale) notFound();
+
+  unstable_setRequestLocale(locale);
+
+  return (
+    <NextIntlClientProvider locale={locale} messages={messages}>
+      {children}
+    </NextIntlClientProvider>
+  );
+};
+
+export default NextIntlProvider;
